+++ model/src/main/scala/org/broadinstitute/dsde/workbench/model/google/GoogleModel.scala
@@ -32,6 +32,10 @@ case class ServiceAccountKey(id: ServiceAccountKeyId,
 // Test IAM types.
 case class IamPermission(value: String) extends ValueObject
 
+// BigQuery types
+case class BigQueryDatasetName(value: String) extends ValueObject
+case class BigQueryTableName(value: String) extends ValueObject
+
 // Storage
 final case class GcsBucketName(value: String) extends ValueObject {
   require(value.length <= 63)
@@ -138,6 +142,8 @@ object GoogleModelJsonSupport {
   }
 
   implicit val GoogleProjectFormat = ValueObjectFormat(GoogleProject)
+  implicit val BigQueryDatasetNameFormat = ValueObjectFormat(BigQueryDatasetName)
+  implicit val BigQueryTableNameFormat = ValueObjectFormat(BigQueryTableName)
 
   implicit val ServiceAccountUniqueIdFormat = ValueObjectFormat(ServiceAccountSubjectId)
   implicit val ServiceAccountNameFormat = ValueObjectFormat(ServiceAccountName)